+++ PHP/Controleur/ScenarioAffichage/AfficheRadio.php
@@ -7,31 +7,10 @@ require('../../Modele/BDD/ConnectionBDD.php');
 $pdo = ConnectionBDD::getInstance();
 $bdd = $pdo::getpdo();
 
+include("../../View/Prof/viewRadio.php");
 
-// Fonction pour afficher les images radio
-function afficherImagesRadio($bdd, $idpatient)
-{
-    // Préparation de la requête
-    $requete = $bdd->prepare("SELECT * FROM radio WHERE idpatient = ?");
-    $requete->execute(array($idpatient));
+require ("../../Modele/Prof/modeleRadio.php");
 
-    // Vérification si des images ont été trouvées
-    if ($requete->rowCount() == 0) {
-        echo "Aucune image trouvée";
-        return;
-    }
+afficheImages($bdd);
 
-    // Affichage des images
-    while ($resultat = $requete->fetch(PDO::FETCH_ASSOC)) {
-        echo "Image Radio<br>";
-        @$imgData = base64_encode($resultat['content']);
-        @$src = 'data:'.$resultat['type'].';base64,'.$imgData;
-        echo $src;
-        echo "<img src='$src'>";
-    }
-}
-
-// Utilisation de la fonction afficherImagesRadio
-afficherImagesRadio($bdd, 63);
 ?>
-
