+++ PHP/View/Prof/viewRadio.php
@@ -1,3 +1,30 @@
-<?php
-//COCO fait le
-?>
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>Radio</title>
+    <meta charset="UTF-8">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
+    <link rel="stylesheet" href="../../View/Style/chat.css" >
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+</head>
+<body>
+<div class="font">
+    <div class="header">
+        <a href="../../Controleur/Accueil/Accueil.php">
+            <img src="../../View/image/logoIFSI.png" width=120 height=80 alt="leLogo" >
+        </a>
+        <h1>Institut de Formation aux Soins Infirmiers (IFSI)</h1>
+        <div class="deconnexion">
+            <a href="../../Controleur/Accueil/Disconnect.php">
+                <img src="../../View/image/Deconnexion.png" class="icone" width="50" height="50" alt="Déconnexion">
+            </a>
+        </div>
+    </div>
+</div>
+<br>
+<button class="button-90" onclick="window.history.back()">Retour</button>
+<br><br>
+
+</body>
+</html>
